+++ packages/login-panel/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.1.0](https://github.com/OBSAU/rtm-kit/compare/@rtm-ui/login-panel@3.0.3...@rtm-ui/login-panel@3.1.0) (2019-01-23)
+
+
+### Features
+
+* **Login-panel:** Added Gdpr checkbox component to login panel ([06a2cdc](https://github.com/OBSAU/rtm-kit/commit/06a2cdc))
+* **Login-panel:** Added Gdpr checkbox component to login panel ([76f04be](https://github.com/OBSAU/rtm-kit/commit/76f04be))
+
+
+
+
+
+## [3.0.3](https://github.com/OBSAU/rtm-kit/compare/@rtm-ui/login-panel@3.0.2...@rtm-ui/login-panel@3.0.3) (2019-01-21)
+
+
+### Bug Fixes
+
+* **how-it-work:** - update package.js ([ecf24dd](https://github.com/OBSAU/rtm-kit/commit/ecf24dd))
+* **how-it-work:** - update unit test and default props ([ef8a222](https://github.com/OBSAU/rtm-kit/commit/ef8a222))
+* **login-panel:** make text inside dangerousHTML ([829f4f4](https://github.com/OBSAU/rtm-kit/commit/829f4f4))
+* **login-panel:** Styling ui and refactor unit test ([1a6070e](https://github.com/OBSAU/rtm-kit/commit/1a6070e))
+
+
+
+
+
 ## [3.0.2](https://github.com/OBSAU/rtm-kit/compare/@rtm-ui/login-panel@3.0.1...@rtm-ui/login-panel@3.0.2) (2019-01-15)
 
 **Note:** Version bump only for package @rtm-ui/login-panel